Move settings file to top-level
This should make it easier for other apps to import settings like so:
`backpack-ui/settings.json`
