+++ package.json
@@ -5,8 +5,8 @@
   "main": "dist/index.js",
   "scripts": {
     "test": "NODE_ENV=test mocha --reporter spec --compilers js:babel-register --recursive spec/",
-    "build": "rm -rf dist && mkdir -p dist/components && babel -d dist src && cp src/settings.json dist/",
-    "ci": "rm -rf dist && mkdir -p dist/components && babel --watch -d dist src && cp src/settings.json dist/",
+    "build": "rm -rf dist && mkdir -p dist/components && babel -d dist src",
+    "ci": "rm -rf dist && mkdir -p dist/components && babel --watch -d dist src",
     "test:ci": "NODE_ENV=test mocha --reporter spec --compilers js:babel-register --watch --recursive spec/"
   },
   "author": "Lonely Planet",
