+++ src/components/button/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import radium from "radium";
 import assign from "object-assign";
-import settings from "../../settings.json";
+import settings from "../../../settings.json";
 import { lighten } from "../../utils/color";
 import { outline } from "../../utils/mixins";
