+++ src/components/contactBar/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import radium from "radium";
 import assign from "object-assign";
-import { color, timing } from "../../settings.json";
+import { color, timing } from "../../../settings.json";
 import Icon from "../icon";
 import { outline } from "../../utils/mixins";
