+++ src/components/flyout/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import radium from "radium";
-import { color, zIndex } from "../../settings.json";
+import { color, zIndex } from "../../../settings.json";
 import { rgb } from "../../utils/color";
 
 const styles = {
