+++ src/components/form/daterange.jsx
@@ -4,7 +4,7 @@ import assign from "object-assign";
 import radium, { Style } from "radium";
 import DateRangePicker from "react-dates/lib/components/DateRangePicker";
 import { END_DATE } from "react-dates/constants";
-import { color, timing, zIndex } from "../../settings.json";
+import { color, timing, zIndex } from "../../../settings.json";
 import { darken, rgb } from "../../utils/color";
 
 const styles = {
