+++ src/components/form/inputGroup.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import radium, { Style } from "radium";
-import { color, media } from "../../settings.json";
+import { color, media } from "../../../settings.json";
 import { darken, rgb } from "../../utils/color";
 import Label from "../form/label";
