+++ src/components/iconButton/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import radium from "radium";
 import assign from "object-assign";
-import settings from "../../settings.json";
+import settings from "../../../settings.json";
 import Icon from "../icon";
 import { darken } from "../../utils/color";
 import { outline } from "../../utils/mixins";
