+++ src/components/location/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import radium from "radium";
-import { color } from "../../settings.json";
+import { color } from "../../../settings.json";
 import MoreLink from "../moreLink";
 import StaticMap from "../staticMap";
 import { blueLink } from "../../utils/mixins";
