+++ src/components/numberList/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import radium from "radium";
-import { color, timing } from "../../settings.json";
+import { color, timing } from "../../../settings.json";
 import NumberMarker from "../numberMarker";
 
 const styles = {
