+++ src/components/paginatorButton/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import radium from "radium";
 import assign from "object-assign";
 import upperFirst from "lodash/upperFirst";
-import settings from "../../settings.json";
+import settings from "../../../settings.json";
 import Icon from "../icon";
 import { rgb } from "../../utils/color";
 import { outline } from "../../utils/mixins";
