+++ src/components/price/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import radium from "radium";
-import { color, media } from "../../settings.json";
+import { color, media } from "../../../settings.json";
 import Bullet from "../decoration/bullet";
 
 const styles = {
