+++ src/components/shareMenu/styles.js
@@ -1,4 +1,4 @@
-import settings from "../../settings.json";
+import settings from "../../../settings.json";
 
 const styles = {
   container: {
