+++ src/components/staticMap/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import radium from "radium";
-import { color, timing } from "../../settings.json";
+import { color, timing } from "../../../settings.json";
 import { span, percentage } from "../../utils/grid";
 
 const styles = {
