+++ src/components/tag/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import radium from "radium";
-import { color, timing } from "../../settings.json";
+import { color, timing } from "../../../settings.json";
 import { rgb } from "../../utils/color";
 
 const tagColor = "#e1eaf0";
