+++ src/components/toolbar/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import radium from "radium";
-import { color, timing } from "../../settings.json";
+import { color, timing } from "../../../settings.json";
 import Icon from "../icon";
 
 const hover = {
